+++ test/test.carbone.js
@@ -57,6 +57,19 @@ describe('Carbone', function () {
     });
   });
 
+  describe(':set', function () {
+    it('should save the result in data, and it can be used later', function (done) {
+      var _xml = '<xml>{d.title:substr(0, 4):set(d.other)} {d.other} </xml>';
+      var _data = {title : 'boo1234'};
+      var _complement = {date : 'today'};
+      carbone.renderXML(_xml, _data, {complement : _complement}, function (err, _xmlBuilt) {
+        helper.assert(err+'', 'null');
+        helper.assert(_xmlBuilt, '<xml> boo1 </xml>');
+        done();
+      });
+    });
+  });
+
   describe('format date', function () {
     afterEach(function (done) {
       carbone.reset();