+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+### v4.5.2
+  - ⭐️ Add the possibility to store values in data with new formatter `:set(d.value)`. Known limits: 
+    - accepts only to store values at the root level of `d`, and only in `d`.
+    - cannot be used in aliases
+
 ### v4.5.1
   - Release October 7th 2022
   - Added for the `:drop(element)` formatter: